client setup #2
