+++ library/src/main/AndroidManifest.xml
@@ -10,7 +10,6 @@
 
     <application>
         <activity android:name=".view.activity.UsersActivity"></activity>
-        <activity android:name=".view.activity.ChatActivity" />
         <activity android:name=".view.activity.AttachmentMediaActivity" />
         <activity android:name=".view.activity.AttachmentDocumentActivity" />
         <activity android:name=".view.activity.AttachmentActivity" />
