+++ library/src/main/java/com/getstream/sdk/chat/interfaces/ChannelEventHandler.java
@@ -3,6 +3,7 @@ package com.getstream.sdk.chat.interfaces;
 import com.getstream.sdk.chat.model.Event;
 
 public interface ChannelEventHandler {
-    void handleEventWSResponse(Event event);
-    void onFailed(String errMsg, int errCode);
+    void handleEventResponse(Event event);
+    void handleConnection();
+    void onConnectionFailed(String errMsg, int errCode);
 }
