+++ library/src/main/java/com/getstream/sdk/chat/interfaces/ChannelListEventHandler.java
@@ -3,5 +3,6 @@ package com.getstream.sdk.chat.interfaces;
 
 public interface ChannelListEventHandler {
     void updateChannels();
+    void handleConnection();
     void onConnectionFailed(String errMsg, int errCode);
 }
