+++ library/src/main/java/com/getstream/sdk/chat/rest/Channel.java
@@ -1,13 +1,12 @@
 package com.getstream.sdk.chat.rest;
 
-import com.getstream.sdk.chat.rest.controller.ChannelRetrofit;
 import com.getstream.sdk.chat.rest.request.SendMessageRequest;
 import com.getstream.sdk.chat.rest.response.MessageResponse;
 
 import retrofit2.Call;
 
 public class Channel {
-    private ChannelRetrofit retrofitClient;
+
     private String type;
     private String id;
     private Client client;
@@ -19,8 +18,8 @@ public class Channel {
 //        retrofitClient = this.client.getRetrofitServiceFactory().create(ChannelRetrofit.class);
     }
 
-    public Call<MessageResponse> sendMessage(Message message){
-        SendMessageRequest request = new SendMessageRequest("", null, null, true, null);
-        return retrofitClient.sendMessage(type, id, client.getApiKey(), request);
-    }
+//    public Call<MessageResponse> sendMessage(Message message){
+//        SendMessageRequest request = new SendMessageRequest("", null, null, true, null);
+//        return retrofitClient.sendMessage(type, id, client.getApiKey(), request);
+//    }
 }
