+++ library/src/main/java/com/getstream/sdk/chat/rest/Client.java
@@ -3,23 +3,12 @@ package com.getstream.sdk.chat.rest;
 import com.getstream.sdk.chat.enums.FilterObject;
 import com.getstream.sdk.chat.enums.QueryOptions;
 import com.getstream.sdk.chat.enums.QuerySort;
-import com.getstream.sdk.chat.interfaces.WSResponseHandler;
-import com.getstream.sdk.chat.model.Event;
-import com.getstream.sdk.chat.rest.controller.ClientRetrofit;
 import com.getstream.sdk.chat.rest.response.GetChannelsResponse;
-import com.getstream.sdk.chat.utils.Global;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.Response;
-import okhttp3.logging.HttpLoggingInterceptor;
-import okio.ByteString;
 import retrofit2.Call;
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
 
 public class Client {
     private ClientRetrofit retrofitClient;
