--- library/src/main/java/com/getstream/sdk/chat/view/activity/BaseActivity.java
@@ -1,40 +1,0 @@
-package com.getstream.sdk.chat.view.activity;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
-
-public class BaseActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-    }
-
-    @Override
-    public void onStart() {
-        super.onStart();
-    }
-
-    @Override
-    public void onStop() {
-        super.onStop();
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-    }
-
-    @Override
-    protected void onPause() {
-        super.onPause();
-        Log.d("BaseActivity:","Background");
-    }
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-    }
-
-}
